requirements.txt
Flask==3.0.3
requests==2.31.0
beautifulsoup4==4.12.3
PySocks==1.7.1
python-telegram-bot==20.7
scikit-learn==1.5.2
joblib==1.3.2
Flask-Login==0.6.3
Werkzeug==3.0.2
APScheduler